+++ src/main/java/com/dongyang/dongpo/domain/member/service/MemberService.java
@@ -1,130 +1,26 @@
 package com.dongyang.dongpo.domain.member.service;
 
-import com.dongyang.dongpo.common.exception.CustomException;
-import com.dongyang.dongpo.common.exception.ErrorCode;
-import com.dongyang.dongpo.common.fileupload.s3.S3Service;
-import com.dongyang.dongpo.domain.auth.dto.*;
+import com.dongyang.dongpo.domain.auth.dto.UserInfo;
 import com.dongyang.dongpo.domain.member.dto.MyPageResponseDto;
 import com.dongyang.dongpo.domain.member.dto.MyPageUpdateRequestDto;
 import com.dongyang.dongpo.domain.member.entity.Member;
-import com.dongyang.dongpo.domain.member.entity.MemberTitle;
-import com.dongyang.dongpo.domain.member.enums.Status;
-import com.dongyang.dongpo.domain.member.repository.MemberRepository;
-import com.dongyang.dongpo.domain.store.service.StoreService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
-@Service
-@Slf4j
-@Transactional
-@RequiredArgsConstructor
-public class MemberService {
+public interface MemberService {
+    Member registerNewMember(UserInfo userInfo);
 
-    @Value("${cloud.aws.s3.bucket-full-url}")
-    private String bucketFullUrl;
+    List<Member> findAll();
 
-    private final MemberRepository memberRepository;
-    private final TitleService titleService;
-    private final StoreService storeService;
-    private final S3Service s3Service;
+    Member findById(Long id);
 
-    // 회원 가입 처리
-    public Member registerNewMember(final UserInfo userInfo) {
-        final Member member = userInfo.toMemberEntity();
-        memberRepository.save(member);
-        titleService.addTitle(member, member.getMainTitle());
-        log.info("Registered Member {} successfully - id: {}", member.getEmail(), member.getId());
-        return member;
-    }
+    Member findByEmail(String email);
 
-    // 사용자 전체 조회
-    public List<Member> findAll() {
-        return memberRepository.findAll();
-    }
+    void setMemberStatusLeave(Member member);
 
-    // id로 사용자 조회
-    public Member findById(final Long id) {
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
-    }
+    boolean validateMemberExistence(String email, String socialId);
 
-    // 이메일로 사용자 조회
-    public Member findByEmail(final String email) {
-        return memberRepository.findByEmail(email)
-                .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
-    }
+    MyPageResponseDto getMemberInfo(Member member);
 
-    // 탈퇴 처리
-    public void setMemberStatusLeave(final Member member) {
-        findByEmail(member.getEmail()).setMemberStatusLeave();
-
-        if (member.getProfilePic() != null && member.getProfilePic().startsWith(bucketFullUrl))
-            s3Service.deleteFile(member.getProfilePic()); // S3에 있는 기존 프로필 사진 삭제
-    }
-
-    // 이미 가입 된 회원인지, 가입 가능한 회원인지 검증
-    public boolean validateMemberExistence(final String email, final String socialId) {
-        // 이미 가입된 회원인지 확인
-        final Optional<Member> existingMemberOpt = memberRepository.findBySocialId(socialId);
-
-        // 이미 가입된 회원인 경우
-        if (existingMemberOpt.isPresent()) {
-            final Member existingMember = existingMemberOpt.get();
-
-            // 탈퇴한 회원인 경우 (탈퇴 시 개인 정보를 모두 삭제하므로 이 검증 과정이 필수는 아님. 더블 체크를 위해 남겨둠.)
-            if (existingMember.getStatus() == Status.LEAVE)
-                throw new CustomException(ErrorCode.MEMBER_ALREADY_LEFT);
-
-            // 모든 조건 통과 시 로그인 처리
-            return true;
-        }
-
-        // 가입 되지 않은 회원일 경우, 이메일 중복 검사
-        if (memberRepository.existsByEmail(email))
-            throw new CustomException(ErrorCode.MEMBER_EMAIL_DUPLICATED);
-
-        return false;
-    }
-
-    public MyPageResponseDto getMemberInfo(final Member member) {
-        return MyPageResponseDto.builder()
-                .nickname(member.getNickname())
-                .profilePic(member.getProfilePic())
-                .mainTitle(member.getMainTitle().getDescription())
-                .registerCount(storeService.getMyRegisteredStoreCount(member))
-                .titleCount(titleService.getMemberTitlesCount(member))
-                .build();
-    }
-
-    public void updateMemberInfo(final Member member, final MyPageUpdateRequestDto myPageUpdateRequestDto) {
-        if (myPageUpdateRequestDto.getNickname().length() > 7) // 문자 수 7자 초과시 예외 발생
-            throw new CustomException(ErrorCode.ARGUMENT_NOT_SATISFIED);
-
-        final Member existingMember = findById(member.getId());
-
-        if (myPageUpdateRequestDto.getProfilePic() != null && !myPageUpdateRequestDto.getProfilePic().isBlank() &&
-                !myPageUpdateRequestDto.getProfilePic().equals(existingMember.getProfilePic())) {
-            if (existingMember.getProfilePic() != null && existingMember.getProfilePic().startsWith(bucketFullUrl)) {
-                s3Service.deleteFile(existingMember.getProfilePic()); // S3에 있는 기존 프로필 사진 삭제
-            }
-            existingMember.updateMemberProfilePic(myPageUpdateRequestDto.getProfilePic());
-            log.info("Member {} - updated profilePic", existingMember.getEmail());
-        }
-        if (myPageUpdateRequestDto.getNickname() != null && !myPageUpdateRequestDto.getNickname().equals(existingMember.getNickname())) {
-            existingMember.updateMemberNickname(myPageUpdateRequestDto.getNickname());
-            log.info("Member {} - updated nickname", existingMember.getEmail());
-        }
-        if (myPageUpdateRequestDto.getNewMainTitle() != null && !myPageUpdateRequestDto.getNewMainTitle().equals(existingMember.getMainTitle())) {
-            MemberTitle memberTitle = titleService.findByMemberAndTitle(existingMember, myPageUpdateRequestDto.getNewMainTitle());
-            existingMember.updateMemberMainTitle(memberTitle.getTitle());
-            log.info("Member {} - updated mainTitle", existingMember.getEmail());
-        }
-    }
-
+    void updateMemberInfo(Member member, MyPageUpdateRequestDto myPageUpdateRequestDto);
 }
